+++ src/controllers/utils/user.messages.ts
@@ -12,10 +12,12 @@ enum ErrorMessagesForUsers {
   errorEnrolling2 = ' es personero y por lo tanto no puede postular.',
   errorNotifying = 'Hubo un error generando la contraseña del usuario.',
   errorVerifying = 'Hubo un error mientras se intentaba verificar el ',
+  errorVerifyingUser = 'Hubo un error verificando al usuario.',
   generic = 'Hubo un error intentando inscribir al usuario',
   studentNotFound = 'El estudiante no se encuentra registrado en el padrón.',
   teacherNotFound = 'El docente no se encuentra registrado en el padrón.',
-  userHasNotMail = 'Ud. no tiene un correo registrado en el padrón electoral.\n Se ha notificado al CEUNI, se comunicarán con Ud. lo antes posible. Disculpe las molestias.'
+  userHasNotMail = 'Ud. no tiene un correo registrado en el padrón electoral.\n Se ha notificado al CEUNI, se comunicarán con Ud. lo antes posible. Disculpe las molestias.',
+  userNotFound = 'Ud. no se encuentra registrado en el padrón electoral.'
 }
 
 enum MessagesForUsers {